+++ careers/careers/vite.config.js
