--- careers/src/components/CareersForm.css
@@ -1,106 +1,0 @@
-.inter-font {
-  font-family: "Inter", sans-serif;
-  font-optical-sizing: auto;
-  font-weight: 100;
-  font-style: normal;
-}
-
-.form-container {
-  width: 80%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 50px auto;
-  padding: 20px;
-  color: black;
-  /* background-color: yellow; */
-}
-
-.form-group {
-  width: 70%;
-  /* background-color : red; */
-  margin-bottom: 15px;
-}
-.entry {
-  display: flex;
-  justify-content: space-between;
-  /* align-items: center; */
-  margin-bottom: 15px;
-}
-  .entry input {
-    width: 40%; /* Adjust width as needed */
-    background-color: #f9f9f9;
-    border: 1px solid #ccc;
-    border-radius: 12px;
-
-  }
-  .entry a{
-    width: 40%; /* Adjust width as needed */
-  }
-  .entry div{
-    width: 40%; /* Adjust width as needed */  
-  }
-  .file-input {
-    display: none;
-  }
-  .country-code{
-    background-color : white;
-    color : black;
-    border : 0px;
-  }
-  .phone-number-input {
-    padding: 5px;
-    border: 1px solid #ccc;
-    border-radius: 0 4px 4px 0;
-    border-left: none;
-    color : black;
-    flex-grow: 1;
-  }
-.input-color{
-  color :black;
-}
-
-
-.button-container {
-  display: flex;
-  justify-content: flex-end; /* Align buttons to the right */
-  gap: 10px; /* Space between buttons */
-  margin-top: 20px; /* Add some space above the buttons */
-  margin-left : auto;
-}
-
-.submit-button {
-  display: flex;
-  height: 30px;
-  width: 150px;
-  justify-content: center;
-  align-items: center;
-  background-color: #ff6600;
-  color: #fff;
-  border: none;
-  border-radius: 12px;
-  cursor: pointer;
-  font-size: 16px; /* Increase font size */
-}
-
-.submit-button:hover {
-  background-color: #e55b00;
-}
-
-.cancel-button {
-  display: flex;
-  height: 30px;
-  width: 150px;
-  justify-content: center;
-  align-items: center;
-  background-color: #cccccc; /* Different background color */
-  color: #000; /* Different text color */
-  border: 1px solid #999; /* Different border */
-  border-radius: 8px; /* Different border radius */
-  cursor: pointer;
-  font-size: 14px; /* Different font size */
-}
-
-.cancel-button:hover {
-  background-color: #bbbbbb; /* Different hover color */
-}
